refactoring for kotlin 1.5 due to deprecated language methods

+++ src/main/kotlin/de/gmuth/ipp/core/IppInputStream.kt
@@ -145,7 +145,7 @@ class IppInputStream(inputStream: InputStream) : DataInputStream(inputStream) {
                             minutes = read(),
                             seconds = read(),
                             deciSeconds = read(),
-                            directionFromUTC = readByte().toChar(),
+                            directionFromUTC = readByte().toInt().toChar(),
                             hoursFromUTC = read(),
                             minutesFromUTC = read()
                     )
